+++ deploy/scripts/pwsh/SAPDeploymentUtilities/Internal/new_saplibrary.ps1
@@ -35,7 +35,7 @@ function New-SAPLibrary {
 Copyright (c) Microsoft Corporation.
 Licensed under the MIT license.
 #>
-    [cmdletbinding()]
+    [cmdletbinding(SupportsShouldProcess)]
     param(
         #Parameter file
         [Parameter(Mandatory = $true)][string]$Parameterfile,
@@ -46,39 +46,44 @@ Licensed under the MIT license.
     Write-Host -ForegroundColor green ""
     Write-Host -ForegroundColor green "Bootstrap the library"
 
-    Add-Content -Path "log.txt" -Value "Bootstrap the library"
-    Add-Content -Path "log.txt" -Value (Get-Date -Format "yyyy-MM-dd HH:mm")
+    $fInfo = Get-ItemProperty -Path $Parameterfile
+    if (!$fInfo.Exists ) {
+        Write-Error ("File " + $Parameterfile + " does not exist")
+        return
+    }
+
+    Add-Content -Path "deployment.log" -Value "Bootstrap the library"
+    Add-Content -Path "deployment.log" -Value (Get-Date -Format "yyyy-MM-dd HH:mm")
     
 
     $mydocuments = [environment]::getfolderpath("mydocuments")
     $filePath = $mydocuments + "\sap_deployment_automation.ini"
-    $iniContent = Get-IniContent $filePath
+    $iniContent = Get-IniContent -Path $filePath
 
     $jsonData = Get-Content -Path $Parameterfile | ConvertFrom-Json
     $region = $jsonData.infrastructure.region
 
-    # Subscription
-    try {
-        $sub = $iniContent[$region]["subscription"] 
-        
-    }
-    catch {
+    # Subscription & repo path
+
+    $sub = $iniContent[$region]["subscription"] 
+    $repo = $iniContent["Common"]["repo"]
+
+    $changed = $false
+
+    if ($null -eq $sub -or "" -eq $sub) {
         $sub = Read-Host -Prompt "Please enter the subscription"
         $iniContent[$region]["subscription"] = $sub
         $changed = $true
-      
     }
 
-    try {
-        $repo = $iniContent["Common"]["repo"]
-    }
-    catch {
+    if ($null -eq $repo -or "" -eq $repo) {
+        $repo = Read-Host -Prompt "Please enter the path to the repository"
         $iniContent["Common"]["repo"] = $repo
         $changed = $true
     }
 
     if ($changed) {
-         Out-IniFile -InputObject $iniContent -FilePath $filePath
+        Out-IniFile -InputObject $iniContent -Path $filePath
     }
 
     $terraform_module_directory = $repo + "\deploy\terraform\bootstrap\sap_library"
@@ -99,17 +104,18 @@ Licensed under the MIT license.
                 $Command = " init -upgrade=true -reconfigure " + $terraform_module_directory
             }
         }
-        else
-        {
-            $ans = Read-Host -Prompt "The system has already been deployed, do you want to redeploy Y/N?"
-            if ("Y" -ne $ans) {
-                return
+        else {
+            if ($PSCmdlet.ShouldProcess($Parameterfile, $DeployerFolderRelativePath)) {
+                $ans = Read-Host -Prompt "The system has already been deployed, do you want to redeploy Y/N?"
+                if ("Y" -ne $ans) {
+                    return
+                }
             }
         }
     }
     
     $Cmd = "terraform $Command"
-    Add-Content -Path "log.txt" -Value $Cmd
+    Add-Content -Path "deployment.log" -Value $Cmd
     & ([ScriptBlock]::Create($Cmd)) 
     if ($LASTEXITCODE -ne 0) {
         throw "Error executing command: $Cmd"
@@ -125,7 +131,7 @@ Licensed under the MIT license.
 
     
     $Cmd = "terraform $Command"
-    Add-Content -Path "log.txt" -Value $Cmd
+    Add-Content -Path "deployment.log" -Value $Cmd
     $planResults = & ([ScriptBlock]::Create($Cmd)) | Out-String 
     
     if ($LASTEXITCODE -ne 0) {
@@ -150,51 +156,54 @@ Licensed under the MIT license.
 
     Write-Host $planResults
 
-    Write-Host -ForegroundColor green "Running apply"
-    if ($DeployerFolderRelativePath -eq "") {
-        $Command = " apply -var-file " + $Parameterfile + " " + $terraform_module_directory
-    }
-    else {
-        $Command = " apply -var-file " + $Parameterfile + " -var deployer_statefile_foldername=" + $DeployerFolderRelativePath + " " + $terraform_module_directory
-    }
+    if ($PSCmdlet.ShouldProcess($Parameterfile, $DeployerFolderRelativePath)) {
+    
+        Write-Host -ForegroundColor green "Running apply"
+        if ($DeployerFolderRelativePath -eq "") {
+            $Command = " apply -var-file " + $Parameterfile + " " + $terraform_module_directory
+        }
+        else {
+            $Command = " apply -var-file " + $Parameterfile + " -var deployer_statefile_foldername=" + $DeployerFolderRelativePath + " " + $terraform_module_directory
+        }
 
-    $Cmd = "terraform $Command"
-    Add-Content -Path "log.txt" -Value $Cmd
-    & ([ScriptBlock]::Create($Cmd))  
-    if ($LASTEXITCODE -ne 0) {
-        throw "Error executing command: $Cmd"
-    }
+        $Cmd = "terraform $Command"
+        Add-Content -Path "deployment.log" -Value $Cmd
+        & ([ScriptBlock]::Create($Cmd))  
+        if ($LASTEXITCODE -ne 0) {
+            throw "Error executing command: $Cmd"
+        }
 
-    New-Item -Path . -Name "backend.tf" -ItemType "file" -Value "terraform {`n  backend ""local"" {}`n}" -Force 
+        New-Item -Path . -Name "backend.tf" -ItemType "file" -Value "terraform {`n  backend ""local"" {}`n}" -Force 
 
-    $Command = " output remote_state_resource_group_name"
-    $Cmd = "terraform $Command"
-    $rgName = & ([ScriptBlock]::Create($Cmd)) | Out-String 
-    if ($LASTEXITCODE -ne 0) {
-        throw "Error executing command: $Cmd"
-    }
-    $iniContent[$region]["REMOTE_STATE_RG"] = $rgName.Replace("""","")
+        $Command = " output remote_state_resource_group_name"
+        $Cmd = "terraform $Command"
+        $rgName = & ([ScriptBlock]::Create($Cmd)) | Out-String 
+        if ($LASTEXITCODE -ne 0) {
+            throw "Error executing command: $Cmd"
+        }
+        $iniContent[$region]["REMOTE_STATE_RG"] = $rgName.Replace("""", "")
 
-    $Command = " output remote_state_storage_account_name"
-    $Cmd = "terraform $Command"
-    $saName = & ([ScriptBlock]::Create($Cmd)) | Out-String 
-    if ($LASTEXITCODE -ne 0) {
-        throw "Error executing command: $Cmd"
-    }
-    $iniContent[$region]["REMOTE_STATE_SA"] = $saName.Replace("""","")
+        $Command = " output remote_state_storage_account_name"
+        $Cmd = "terraform $Command"
+        $saName = & ([ScriptBlock]::Create($Cmd)) | Out-String 
+        if ($LASTEXITCODE -ne 0) {
+            throw "Error executing command: $Cmd"
+        }
+        $iniContent[$region]["REMOTE_STATE_SA"] = $saName.Replace("""", "")
 
-    $Command = " output tfstate_resource_id"
-    $Cmd = "terraform $Command"
-    $tfstate_resource_id = & ([ScriptBlock]::Create($Cmd)) | Out-String 
-    if ($LASTEXITCODE -ne 0) {
-        throw "Error executing command: $Cmd"
-    }
-    $iniContent[$region]["tfstate_resource_id"] = $tfstate_resource_id
+        $Command = " output tfstate_resource_id"
+        $Cmd = "terraform $Command"
+        $tfstate_resource_id = & ([ScriptBlock]::Create($Cmd)) | Out-String 
+        if ($LASTEXITCODE -ne 0) {
+            throw "Error executing command: $Cmd"
+        }
+        $iniContent[$region]["tfstate_resource_id"] = $tfstate_resource_id
 
-    Out-IniFile -InputObject $iniContent -FilePath $filePath
+        Out-IniFile -InputObject $iniContent -Path $filePath
 
-    if (Test-Path ".\backend.tf" -PathType Leaf) {
-        Remove-Item -Path ".\backend.tf" -Force 
+        if (Test-Path ".\backend.tf" -PathType Leaf) {
+            Remove-Item -Path ".\backend.tf" -Force 
+        }
     }
 
 }
